Fix BehaviourTree, aggiunti const, eliminata classe non utilizzata Selector
https://github.com/ranghetto/Qovid/issues/33

+++ src/AI/InfectActorsInRange.cpp
@@ -1,8 +1,8 @@
 #include "InfectActorsInRange.h"
 #include "../Controllers/Simulation.h"
 
-InfectActorsInRange::InfectActorsInRange(const Actor &actor, int range,
-                                         float infectRateo)
+InfectActorsInRange::InfectActorsInRange(const Actor &actor, uint range,
+                                         uint infectRateo)
     : actor_(actor), range_(range), infectRateo_(infectRateo) {}
 
 NodeState InfectActorsInRange::evaluate() {
